label encoding added in loso_lstm

+++ src/datasets/sequence_dataset.py
@@ -25,8 +25,9 @@ class LSTMDataset(Dataset):
         # Return one windowed sequence and its label 
         x_seq = self.features[idx: idx + self.window_size]
         y = self.labels[idx + self.window_size - 1] # timestep label
-        return torch.tensor(x_seq, dtype=torch.float32), torch.tensor(y) 
-    
+        
+        return torch.tensor(x_seq, dtype=torch.float32), torch.tensor(y, dtype=torch.long)
+
 
 # === WHAT DOES THIS DO? ===
 # For a 10-row window, returns:
